First attempt at 64 bit compilation!
filesystem, hash, pefile, tools, meta, toolkit:
* Fixed unsigned int / size_t mismatches.

color:
* Added alpha argument when initializing by colorname.

math_sse, toolkit:
* #ifdef'd out x86 unique code...

tools:
* Moved templated number formatting to printf.h
* Fixed incorrect formatting to limit length of string printf.
* Reworked `CEnumBase::_RenderFlags()` to iterate through Index() calls instead of bits. This allows ES_INT to write all remaining bits as one number.
* Reversed ending with ellipsis on overflow back to simply truncating.
* Explicit buffer initialization of va_buf by construction with an integer as argument (which is ignored).
* Renamed member functions in va_buf for consistency.
* Added va_buf::format() to format any number with correct format string.
* Added cast_vabuf() as a glue helper to safely cast a char array to va_buf.
* Added va_format() convenience helper.
* Updated unittests to be a tad more bareable.

+++ public/filesystem/file.h
@@ -123,7 +123,7 @@ inline void file::rewind()
 {
 	seek( 0, BEGIN );
 }
-inline char* file::dump( unsigned int& size ) const
+inline char* file::dump( size_t& size ) const
 {
 	// Read a complete file & dump to allocated memory, arbitrary 10MB limit
 	// Has to be inline so the proper memory alloc is used, don't forget to ::free() when you're done
